# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Smart Web Scraper API
  description: 지능형 웹 데이터 수집 플랫폼 — 5단계 적응형 파이프라인 + Self-Healing
  version: 1.0.0
  contact:
    name: Smart Web Scraper
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/scrape/now:
    post:
      tags: [Scrape]
      summary: 즉시 수집 트리거
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url_id]
              properties:
                url_id: { type: string, format: uuid }
                pipeline_override: { type: string, enum: [api, http, stealth, ai, proxy] }
      responses:
        "202":
          description: 수집 태스크 시작됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id: { type: string }
                  status: { type: string }
  /api/v1/scrape/bulk:
    post:
      tags: [Scrape]
      summary: 복수 URL 일괄 수집
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url_ids]
              properties:
                url_ids:
                  type: array
                  items: { type: string, format: uuid }
                  maxItems: 50
      responses:
        "202": { description: 일괄 수집 시작됨 }
  /api/v1/scrape/{task_id}/status:
    get:
      tags: [Scrape]
      summary: 태스크 상태 조회
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id: { type: string }
                  status: { type: string, enum: [pending, running, completed, failed] }
                  result: { $ref: "#/components/schemas/ScrapeResult" }
  /api/v1/scrape/{task_id}/cancel:
    post:
      tags: [Scrape]
      summary: 태스크 취소
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 취소됨 }

  # ─── Schedules ────────────────────────────────────────────
  /api/v1/schedules:
    get:
      tags: [Schedules]
      summary: 스케줄 목록
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScheduleResponse" }

    post:
      tags: [Schedules]
      summary: 스케줄 등록
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleCreate" }
      responses:
        "201": { description: 등록 성공 }
  /api/v1/schedules/{id}:
    put:
      tags: [Schedules]
      summary: 스케줄 수정
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: 수정 성공 }

    delete:
      tags: [Schedules]
      summary: 스케줄 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 삭제 성공 }
  /api/v1/schedules/{id}/pause:
    post:
      tags: [Schedules]
      summary: 스케줄 일시정지
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: 일시정지됨 }
  /api/v1/schedules/{id}/resume:
    post:
      tags: [Schedules]
      summary: 스케줄 재개
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: 재개됨 }
  /api/v1/schedules/{id}/run:
    post:
      tags: [Schedules]
      summary: 스케줄 즉시 실행
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "202": { description: 실행 시작됨 }

  # ─── Dashboard ────────────────────────────────────────────
...
components:
  schemas:
...
    ScrapeResult:
      type: object
      properties:
        url: { type: string }
        success: { type: boolean }
        pipeline_name: { type: string }
        data: { type: object }
        duration_ms: { type: integer }
        error_type: { type: string }
        error_message: { type: string }
        healing_applied: { type: boolean }
    ScheduleCreate:
      type: object
      required: [url_id, cron_expression]
      properties:
        url_id: { type: string, format: uuid }
        schedule_type: { type: string, enum: [hourly, daily, weekly, monthly, custom] }
        cron_expression: { type: string, example: "0 9 * * *" }
        timezone: { type: string, default: Asia/Seoul }
    ScheduleResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        url_id: { type: string, format: uuid }
        schedule_type: { type: string }
        cron_expression: { type: string }
        is_active: { type: boolean }
        next_run_at: { type: string, format: date-time }
        run_count: { type: integer }
        success_count: { type: integer }
...